# src/main/java/com/amanda/estoqueAPI/model/Produto.java
package com.amanda.estoqueAPI.model;

import com.amanda.estoqueAPI.model.Enum.TipoProduto;
import jakarta.persistence.*;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;
import org.springframework.data.jpa.domain.support.AuditingEntityListener;

import java.math.BigDecimal;
import java.util.List;
import java.util.UUID;

@Entity
@Table (name = "produto", schema = "public")
@Getter
@Setter
@EntityListeners(AuditingEntityListener.class)
public class Produto {

    @Id
    @Column (name = "id")
    @GeneratedValue (strategy = GenerationType.UUID)
    private UUID id;

    @Column (name = "nome", length = 100, nullable = false)
    private String nome;

    @Column (name = "tamanho", length = 20, nullable = false)
    private String tamanho;

    @Column (name = "quantidade", nullable = false)
    private Integer quantidade;

    @Column (name = "preco", precision = 18, scale = 2, nullable = false)
    private BigDecimal preco;

//    @OneToMany (mappedBy = "produto")
//    private List<ItemVenda> itens;

//    @Column (name = "cor", length = 20, nullable = true )
//    private String cor;
//
//    @Column (name = "estampa", length = 100, nullable = true)
//    private String estampa;
//
//    @Column (name = "descricao", length = 100, nullable = true)
//    private String descricao;

//    @Enumerated (EnumType.STRING)
//    @Column (name = "tipo_produto")
//    private TipoProduto tipoProduto;


    public String toString (){
        return "Autor{" +
                "id=" + id +
                ", nome='" + nome + '\'' +
                ", tamanho=" + tamanho +
                ", quantidade=" + quantidade + '\'' +
                ", preço=" + preco + '\'' +
                '}';
    }

}